Continued work on survey controller/model
Approach to data has changed: we are now building with a reduced scope,
where a survey constitutes a single question rather than multiple questions

refactored survey schema to reflect that survey was now equivalent to question

started building virtual methods to handle processing of results. Next step
is to test these virtual methods in the console. Not sure how to do this,
working with Pinaro to figure it out.