Merge prep
Realized that my efforts to work with adding results to a survey are
moot without testing whether or not it is possible to CRUD from the
front end. Commented out related code and will start working on the front
end.

Going to be using an old commit from my second project to create basic
crud functionality.

Worth noting that I am having some weirdness with the schema, where changes
made to it are not reflected in mongoDB. Unsure of why this is.