Out from the quagmire
Courtesy of Scott, realized that I can crack open the black box of my
server via curl and console logs. This allowed me to hash out the logical
handling of the update method. As things stand, I am near the end of the
road on getting this to curl successfully. All that stands in my way is
an improperly wired save function.