Close to the end of building update
Only hurdle in the way is that I cannot save to the database. updatedQuestions
is undefined, causing an error. Unsure of why this is the case, seemed
to work for both wdi-16 and is basically a hard copy of the docs method.

Going to do some testing on Pinaro's discriminate index code until a consultant
answers the question.