OUT OF THE DARKNESS
Finally was able to resolve the problem of updating my questions. It turns
out that simply using the mongoose update method allowed for resolution,
where treating it like normal javascript objects failed.

With this done, the core functionality on the back end is now almost complete.
All that needs to be done now is applying the discriminate index to questions
and refactoring userSurveys to fit the updated schema